photoShareBackend/requirements.txt
asgiref==3.3.1
cffi==1.14.5
cryptography==3.4.6
Django==3.1.7
django-appconf==1.0.4
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-imagekit==4.0.2
django-rest-framework==0.1.0
django-rest-knox==4.1.0
djangorestframework==3.12.2
fake-factory==9999.9.9
Faker==6.5.0
mysqlclient==2.0.3
pilkit==2.0
Pillow==8.1.2
pycparser==2.20
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
